src/backend/requirements.txt
gunicorn==21.2.0
uvicorn==0.23.0
pydantic==2.10.5
aiohttp==3.8.5
python-dotenv==1.0.0

sqlalchemy==2.0.0
asyncpg==0.28.0
alembic==1.11.0
sqlalchemy-utils==0.41.0
sqlalchemy-timescaledb==0.4.1

redis==4.6.0
aioredis==1.3.1

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0

pynvml==11.5.0
docker==6.1.0
kubernetes==27.2.0

prometheus-client==0.17.0
python-json-logger==2.0.7
json-logging==1.3.0
structlog==23.1.0
elasticsearch-logger==0.0.2
sentry-sdk==1.28.1

stripe==5.4.0
fastapi==0.100.0
fastapi-limiter==0.1.5
fastapi-cache==0.1.0
tenacity==8.0.0
cachetools==5.0.0
circuitbreaker==1.4.0
loguru==0.7.0

boto3==1.35.96
botocore==1.35.96
minio==7.1.0

pyroute2==0.7.3

websockets==11.0.0
jsonschema==4.0.0
iso4217==1.12.20240625

pytest==7.4.0
pytest-asyncio==0.21.0
httpx==0.24.0
freezegun==1.2.0
faker==19.3.0
black==23.7.0
flake8==6.1.0
mypy==1.4.0
isort==5.12.0
psycopg2==2.9.10

numpy==2.0.2
numba==0.60.0
pycuda==2024.1.2

ratelimit==2.2.1
